+++ app/(app)/beans/new/page.tsx
@@ -1,4 +1,4 @@
-import { PlaceholderPage } from "@/components/shared/placeholder-page"
+import { PlaceholderPage } from "@/components/shared/PlaceholderPage"
 import { beansRoutes } from "@/features/beans"
 import { requireRoasterMembership } from "@/server/auth"
