+++ app/(app)/likes/page.tsx
@@ -1,4 +1,4 @@
-import { PlaceholderPage } from "@/components/shared/placeholder-page"
+import { PlaceholderPage } from "@/components/shared/PlaceholderPage"
 import { likesRoutes } from "@/features/likes"
 
 export default function LikesPage() {
