+++ app/(app)/offers/[id]/page.tsx
@@ -1,4 +1,4 @@
-import { PlaceholderPage } from "@/components/shared/placeholder-page"
+import { PlaceholderPage } from "@/components/shared/PlaceholderPage"
 import { offersRoutes } from "@/features/offers"
 
 type OfferPageProps = Readonly<{
