+++ app/(app)/search/roasters/page.tsx
@@ -1,4 +1,4 @@
-import { PlaceholderPage } from "@/components/shared/placeholder-page"
+import { PlaceholderPage } from "@/components/shared/PlaceholderPage"
 import { searchRoutes } from "@/features/search"
 
 export default function SearchRoastersPage() {
