+++ app/(app)/users/[id]/page.tsx
@@ -1,4 +1,4 @@
-import { PlaceholderPage } from "@/components/shared/placeholder-page"
+import { PlaceholderPage } from "@/components/shared/PlaceholderPage"
 import { usersRoutes } from "@/features/users"
 
 type UserPageProps = Readonly<{
