+++ app/(app)/users/password/page.tsx
@@ -1,4 +1,4 @@
-import { PlaceholderPage } from "@/components/shared/placeholder-page"
+import { PlaceholderPage } from "@/components/shared/PlaceholderPage"
 import { usersRoutes } from "@/features/users"
 
 export default function UserPasswordPage() {
