+++ app/(app)/wants/[id]/page.tsx
@@ -1,4 +1,4 @@
-import { PlaceholderPage } from "@/components/shared/placeholder-page"
+import { PlaceholderPage } from "@/components/shared/PlaceholderPage"
 import { wantsRoutes } from "@/features/wants"
 
 type WantPageProps = Readonly<{
